% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_group.R
\name{central_group}
\alias{central_group}
\alias{median_group}
\alias{mean_group}
\title{central_group}
\usage{
central_group(
  data,
  group,
  quanti_exp,
  type,
  group.fill = NULL,
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.group = TRUE,
  na.rm.facet = TRUE,
  total = TRUE,
  reorder = FALSE,
  show_ci = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  show_labs = TRUE,
  total_name = NULL,
  digits = 0,
  unit = "",
  dec = NULL,
  col = NULL,
  pal = "OBSS_Spring",
  direction = 1,
  desaturate = 0,
  lighten = 0,
  darken = 0,
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  wrap_width_leg = 25,
  legend_ncol = 4,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  lang = "fr",
  theme = "fonctionr",
  coef_font = 1,
  export_path = NULL
)

median_group(..., type = "median")

mean_group(..., type = "mean")
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{group}{A variable defining groups to be compared.}

\item{quanti_exp}{An expression defining the quantitative variable from which the mean/median is computed. Notice that if any observations with \code{NA} in at least one of the variable in \code{quanti_exp} are excluded for the computation of the indicators.}

\item{type}{\code{"mean"} to compute mean by group ; \code{"median"} to compute median by group.}

\item{group.fill}{A variable defining a second variable of groups to be compared.}

\item{facet}{A variable defining the faceting group.}

\item{filter_exp}{An expression filtering the data, preserving the design. Notice that \code{filter_exp} works as \code{\link[srvyr:dplyr_single]{srvyr::filter()}}: it excludes observations for which \code{filter_exp} results into \code{NA}. It is often the case when \code{NA} is present on one of the filter variables.}

\item{...}{All options possible in \code{\link[srvyr:as_survey_design]{srvyr::as_survey_design()}}.}

\item{na.rm.group}{\code{TRUE} if you want to remove observations with \code{NA} on the group and the \code{group.fill} variables. \code{FALSE} if you want to create a group with the \code{NA} values for the group variable and a \code{group.fill} with the \code{NA} values for the \code{group.fill} variable. Default is \code{TRUE}.}

\item{na.rm.facet}{\code{TRUE} if you want to remove observations with \code{NA} on the facet variable. \code{FALSE} if you want to create a facet with the \code{NA} values for the facet variable. Default is \code{TRUE}.}

\item{total}{\code{TRUE} if you want to compute a total, \code{FALSE} if you don't. The default is \code{TRUE}.}

\item{reorder}{\code{TRUE} if you want to reorder the groups according to the mean/median. \code{NA} value, if \code{na.rm.group = FALSE}, is not included in the reorder. In case of facets, the groups are reordered based on each median group. Default is \code{FALSE}.}

\item{show_ci}{\code{TRUE} if you want to show the error bars on the graphic. \code{FALSE} if you don't want to show the error bars. Default is \code{TRUE}.}

\item{show_n}{\code{TRUE} if you want to show on the graphic the number of observations in the sample in each group. \code{FALSE} if you don't want to show this number. Default is \code{FALSE}.}

\item{show_value}{\code{TRUE} if you want to show the mean/median in each group on the graphic. \code{FALSE} if you don't want to show the mean/median. Default is \code{TRUE}.}

\item{show_labs}{\code{TRUE} if you want to show axes and legend (in case of a \code{group.fill}) labels. \code{FALSE} if you don't want to show any labels on axes and legend. Default is \code{TRUE}.}

\item{total_name}{Name of the total displayed on the graphic. Default is \code{"Total"} in French and in English and \code{"Totaal"} in Dutch.}

\item{digits}{Number of decimal places displayed on the values labels on the graphic. Default is \code{0}.}

\item{unit}{Unit displayed on the graphic. Default is none (\code{""}).}

\item{dec}{Decimal mark displayed on the graphic. Default depends on lang: \code{","} for fr and nl ; \code{"."} for en.}

\item{col}{Color of the bars if there is no \code{group.fill}. \code{col} must be a R color or an hexadecimal color code. Default color used depends on type : \code{"deeppink3"} for mean and \code{"mediumorchid3"} for median. The colors of total and \code{NA} group (in case of \code{na.rm.group = FALSE}) are always \code{"grey40"} and \code{"grey"}. If there is a \code{group.fill}, \code{col} has no effect and \code{pal} argument should be used instead.}

\item{pal}{Colors of the bars if there is a \code{group.fill}. \code{pal} must be vector of R colors or hexadecimal colors or a palette from packages MetBrewer or PrettyCols or a palette from fonctionr. The color of missing values for \code{group.fill} (in case of \code{na.rm.group = FALSE}) and for the total are always \code{"grey"} and \code{"grey40"}. If there is no \code{group.fill}, \code{pal} has no effect and \code{col} argument should be used instead.}

\item{direction}{Direction of the palette color. Default is \code{1}. The opposite direction is \code{-1}. If there is no \code{group.fill}, this argument has no effect.}

\item{desaturate}{Numeric specifying the amount of desaturation where \code{1} corresponds to complete desaturation (no colors, grey layers only), \code{0} to no desaturation, and values in between to partial desaturation. Default is \code{0}. It affects only the palette (\code{pal}, if there is a second group) and not the monocolor (\code{col}, if there is no second group). See \link[colorspace:desaturate]{colorspace::desaturate} function from colorspace package for details. If desaturate and lighten/darken arguments are used, lighten/darken is applied in a second time (i.e. on the color transformed by desaturate).}

\item{lighten}{Numeric specifying the amount of lightening. Negative numbers cause darkening. Value shoud be ranged between \code{-1} (black) and \code{1} (white). Default is \code{0}. It doesn't affect the color of \code{NA} (in case of \code{na.rm.group = FALSE}). It affects only the palette (\code{pal}, if there is a second group) and not the monocolor (\code{col}, if there is no second group). See \link[colorspace:desaturate]{colorspace::desaturate} for details. If both argument ligthen and darken are used (not advised), darken is applied in a second time (i.e. on the color transformed by lighten).}

\item{darken}{Numeric specifying the amount of lightening. Negative numbers cause lightening. Value shoud be ranged between \code{-1} (white) and \code{1} (black). Default is \code{0}. It doesn't affect the color of \code{NA} (in case of \code{na.rm.group = FALSE}). It affects only the palette (\code{pal}, if there is a second group) and not the monocolor (\code{col}, if there is no second group). See \link[colorspace:desaturate]{colorspace::desaturate} for details. If both argument ligthen and darken are used (not advised), darken is applied in a second time (i.e. on the color transformed by lighten).}

\item{dodge}{Width of the bars. Default is \code{0.9} to let a small space between bars. A value of \code{1} leads to no space betweens bars. Values higher than \code{1} are not advised because they cause an overlaping of the bars. \code{dodge} doesn't affect the spaces between second groups (\code{group.fill}). There is always no space between second groups.}

\item{font}{Font used in the graphic. See \code{load_and_active_fonts()} for available fonts. Default is \code{"Roboto"}.}

\item{wrap_width_y}{Number of characters before going to the line for the labels of the groups. Default is \code{25}.}

\item{wrap_width_leg}{Number of characters before going to the line for the labels of the \code{group.fill}. Default is \code{25}.}

\item{legend_ncol}{Number of columns in the legend. Default is \code{4}.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As \code{\link[ggplot2:coord_flip]{ggplot2::coord_flip()}} is used in the graphic, \code{xlab} refers to the x label on the graphic, after the \code{\link[ggplot2:coord_flip]{ggplot2::coord_flip()}}, and not to the x variable in the data. Default (\code{xlab = NULL}) displays, for \code{type = "mean"}, "Moyenne :" (if \code{lang = "fr"}), "Mean:" (if \code{lang = "en"}) or "Gemiddelde:" (if \code{lang = "nl"}), or, for \code{type = "median"}, "Médiane :" (if \code{lang = "fr"}), "Median:" (if \code{lang = "en"}) or "Mediaan:" (if \code{lang = "nl"}), followed by the \code{quanti_exp} argument. To show no X label, use \code{xlab = ""}.}

\item{ylab}{Y label on the graphic. As \code{\link[ggplot2:coord_flip]{ggplot2::coord_flip()}} is used in the graphic, \code{ylab} refers to the y label on the graphic, after the \code{\link[ggplot2:coord_flip]{ggplot2::coord_flip()}}, and not to the y variable in the data. Default (\code{ylab = NULL}) displays the name of the \code{group} variable. To show no Y label, use \code{ylab = ""}.}

\item{legend_lab}{Legend (fill) label on the graphic. If \code{legend_lab = NULL}, legend label on the graphic will be \code{group.fill}. To show no legend label, use \code{legend_lab = ""}.}

\item{caption}{Caption of the graphic. This caption goes under de default caption showing the result of the statistical test. There is no way of not showing the result of the chi-square test as a caption.}

\item{lang}{Language of the indications on the graphic. Possibilities are \code{"fr"} (french), \code{"nl"} (dutch) and \code{"en"} (english). Default is \code{"fr"}.}

\item{theme}{Theme of the graphic. Default is \code{"fonctionr"}. \code{"IWEPS"} adds y axis lines and ticks. \code{NULL} uses the default grey ggplot2 theme.}

\item{coef_font}{A multiplier factor for font size of all fonts on the graphic. Default is \code{1}. Usefull when exporting the graphic for a publication (e.g. in a Quarto document).}

\item{export_path}{Path to export the results in an xlsx file. The file includes three (without \code{group.fill}) or two sheets (with a \code{group.fill}): the table, the graphic and the statistical test result.}
}
\value{
A list that contains a table, a ggplot graphic and, in most cases, a statistical test.
}
\description{
Function to compare means or medians among different groups based on complex survey data. It produces a list containing a table, including the confidence intervals of the indicators, a ready-to-be published ggplot graphic and a statistical test.

In case of mean comparison, the statistical test is a Wald test (using \code{\link[survey:regTermTest]{survey::regTermTest()}}). In case of median comparison the statistical test is a Kruskal Wallis test (using \code{\link[survey:svyranktest]{survey::svyranktest()}}). The confidence intervals and the statistical test are taking into account the complex survey design. In case of facets, the statistical test is computed on the total means or medians between facets (and not within facets). In case of second group (\code{group.fill}), no statistical test is computed.

Exporting the results to an Excell file is possible.
}
\examples{
# Loading of data
data(eusilc, package = "laeken")

# Creation of age categories
eusilc$age_cat <- cut(eusilc$age,
breaks = 6,
include.lowest = TRUE)

# Calculation of income means by age category with fonctionr, taking sample design into account
eusilc_mean <- mean_group(
  eusilc,
  group = age_cat,
  quanti_exp = eqIncome / 12,
  strata = db040,
  ids = db030,
  weight = rb050,
  title = "Mean of equivalised income in household by age of individuals",
  subtitle = "Example with austrian SILC data from 'laeken' package"
)

# Results in graph form
eusilc_mean$graph

# Results in table format
eusilc_mean$tab
}
