% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_srvyr}
\alias{convert_to_srvyr}
\title{Fonction pour convertir en objet srvyr}
\usage{
convert_to_srvyr(data, ...)
}
\arguments{
\item{data}{un dataframe, un objet srvyr ou survey}

\item{...}{}
}
\description{
Fonction pour convertir en objet srvyr
}
