% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{crosstable}
\alias{crosstable}
\title{crosstable : fonction pour calculer facilement des profils lignes par groupe}
\usage{
crosstable(
  data,
  group,
  var_distrib,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  caption = NULL,
  show_prop = TRUE,
  unit = "",
  digits = 0,
  dodge = 0.9,
  pretty_pal = "Hokusai1",
  direction = 1,
  wrap_width = 25,
  na.rm = T
)
}
\arguments{
\item{na.rm}{}
}
\description{
crosstable : fonction pour calculer facilement des profils lignes par groupe
}
