% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_excel}
\alias{export_excel}
\title{Fonction pour exporter en excel les résultats des fonctions de fonctionr}
\usage{
export_excel(
  tab_excel = tab_excel,
  graph = graph,
  test_stat_excel = test_stat_excel,
  facet_null = NULL,
  export_path = export_path,
  percent_fm = NULL,
  fgFill = fgFill,
  bivariate = NULL
)
}
\arguments{
\item{fgFill}{}
}
\description{
Fonction pour exporter en excel les résultats des fonctions de fonctionr
}
