% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctionr_options.R
\name{fonctionr_options}
\alias{fonctionr_options}
\title{fonctionr_options}
\usage{
fonctionr_options(
  na.rm.group = NULL,
  na.rm.facet = NULL,
  na.prop = NULL,
  na.vars = NULL,
  na.rm.var = NULL,
  probs = NULL,
  total = NULL,
  prop_method = NULL,
  quantiles = NULL,
  moustache_probs = NULL,
  bw = NULL,
  resolution = NULL,
  height = NULL,
  limits = NULL,
  reorder = NULL,
  position = NULL,
  show_ci = NULL,
  show_mid_point = NULL,
  show_mid_line = NULL,
  show_ci_errorbar = NULL,
  show_ci_lines = NULL,
  show_ci_area = NULL,
  show_quant_lines = NULL,
  show_moustache = NULL,
  show_n = NULL,
  show_value = NULL,
  show_labs = NULL,
  total_name = NULL,
  scale = NULL,
  digits = NULL,
  unit = NULL,
  dec = NULL,
  col = NULL,
  pal = NULL,
  direction = NULL,
  desaturate = NULL,
  lighten = NULL,
  darken = NULL,
  col_density = NULL,
  col_moustache = NULL,
  col_border = NULL,
  alpha = NULL,
  dodge = NULL,
  font = NULL,
  wrap_width_y = NULL,
  wrap_width_leg = NULL,
  legend_ncol = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  lang = NULL,
  theme = NULL,
  coef_font = NULL,
  parallel = NULL,
  erase_all = FALSE
)
}
\arguments{
\item{na.rm.group}{\code{na.rm.group} argument.}

\item{na.rm.facet}{\code{na.rm.facet} argument.}

\item{na.prop}{\code{na.prop} argument.}

\item{na.vars}{\code{na.vars} argument.}

\item{na.rm.var}{\code{na.rm.var} argument.}

\item{probs}{\code{probs} argument.}

\item{total}{\code{total} argument.}

\item{prop_method}{\code{prop_method} argument.}

\item{quantiles}{\code{quantiles} argument.}

\item{moustache_probs}{\code{moustache_probs} argument.}

\item{bw}{\code{bw} argument.}

\item{resolution}{\code{resolution} argument.}

\item{height}{\code{height} argument.}

\item{limits}{\code{limits} argument.}

\item{reorder}{\code{reorder} argument.}

\item{position}{\code{position} argument.}

\item{show_ci}{\code{show_ci} argument.}

\item{show_mid_point}{\code{show_mid_point} argument.}

\item{show_mid_line}{\code{show_mid_line} argument.}

\item{show_ci_errorbar}{\code{show_ci_errorbar} argument.}

\item{show_ci_lines}{\code{show_ci_lines} argument.}

\item{show_ci_area}{\code{show_ci_area} argument.}

\item{show_quant_lines}{\code{show_quant_lines} argument.}

\item{show_moustache}{\code{show_moustache} argument.}

\item{show_n}{\code{show_n} argument.}

\item{show_value}{\code{show_value} argument.}

\item{show_labs}{\code{show_labs} argument.}

\item{total_name}{\code{total_name} argument.}

\item{scale}{\code{scale} argument.}

\item{digits}{\code{digits} argument.}

\item{unit}{\code{unit} argument.}

\item{dec}{\code{dec} argument.}

\item{col}{\code{col} argument.}

\item{pal}{\code{pal} argument.}

\item{direction}{\code{direction} argument.}

\item{desaturate}{\code{desaturate} argument.}

\item{lighten}{\code{lighten} argument.}

\item{darken}{\code{darken} argument.}

\item{col_density}{\code{col_density} argument.}

\item{col_moustache}{\code{col_moustache} argument.}

\item{col_border}{\code{col_border} argument.}

\item{alpha}{\code{alpha} argument.}

\item{dodge}{\code{dodge} argument.}

\item{font}{\code{font} argument.}

\item{wrap_width_y}{\code{wrap_width_y} argument.}

\item{wrap_width_leg}{\code{wrap_width_leg} argument.}

\item{legend_ncol}{\code{legend_ncol} argument.}

\item{title}{\code{title} argument.}

\item{subtitle}{\code{subtitle} argument.}

\item{xlab}{\code{xlab} argument.}

\item{ylab}{\code{ylab} argument.}

\item{legend_lab}{\code{legend_lab} argument.}

\item{caption}{\code{caption} argument.}

\item{lang}{\code{lang} argument.}

\item{theme}{\code{theme} argument.}

\item{coef_font}{\code{coef_font} argument.}

\item{parallel}{\code{parallel} argument.}

\item{erase_all}{\code{TRUE} erases all the options. Default is \code{FALSE}.}
}
\value{
No return value, called for side effects.
}
\description{
Function to set global options for fonctionr. The arguments defined in the options are only active if the user has not manually specified a value for those arguments within the various functions. Arguments may be shared by multiple functions (if they have the same name) or specific to certain functions.
}
\examples{
# We set global settings
fonctionr_options(coef_font = 1.5, col = "magenta", caption = "Beautiful caption")

# Loading of data
data(eusilc, package = "laeken")

# Recoding eusilc$pl030 into eusilc$pl030_rec
eusilc$pl030_rec <- NA
eusilc$pl030_rec[eusilc$pl030 == "1"] <- "Working full time"
eusilc$pl030_rec[eusilc$pl030 == "2"] <- "Working part time"
eusilc$pl030_rec[eusilc$pl030 == "3"] <- "Unemployed"
eusilc$pl030_rec[eusilc$pl030 == "4"] <- "Student"
eusilc$pl030_rec[eusilc$pl030 == "5"] <- "Retired"
eusilc$pl030_rec[eusilc$pl030 == "6"] <- "Permanently disabled"
eusilc$pl030_rec[eusilc$pl030 == "7"] <- "Fulfilling domestic tasks"

# Computation, taking sample design into account
eusilc_prop <- prop_group(
  eusilc,
  group = pl030_rec,
  prop_exp = py090n > 0,
  weight = rb050,
  title = "\% of ind. receiving unemployment benefits in their hh",
  subtitle = "Example with austrian SILC data from 'laeken' package"
)

# Results in graph form
eusilc_prop$graph

# We set back settings to default
fonctionr_options(erase_all = TRUE)
}
