% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_prop_group.R
\name{n_prop_group}
\alias{n_prop_group}
\title{n_prop_group}
\usage{
n_prop_group(
  data,
  group,
  prop_bin_col,
  facet_var = NULL,
  filter_exp = NULL,
  prop_method = "beta",
  ...,
  unit = "\%",
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  scale = 100,
  digits = 0,
  show_labs = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  dodge = 0.9,
  pretty_pal = "Hokusai1",
  direction = 1,
  error_bar = T,
  na.rm.group = T,
  total_name = "Total",
  font = "Roboto",
  wrap_width = 25
)
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{group}{A variable defining groups be compared.}

\item{prop_bin_col}{A vector containing names of the binarized variables on which to compute the proportions}

\item{...}{All options possible in as_survey_design in srvyr package.}
}
\description{
n_prop_group
}
