% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_survey.R
\name{pivot_longer_survey}
\alias{pivot_longer_survey}
\title{pivot_longer_survey}
\usage{
pivot_longer_survey(data, n_groups)
}
\arguments{
\item{data}{A dataframe with Agregated data to pivot}

\item{n_groups}{Number of groups by which data have been agregated}
}
\value{
A dataframe
}
\description{
Function to pivot from wide to long agregated data by group produced with srvyr::summarise
}
