% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quali_distrib.R
\name{quali_distrib}
\alias{quali_distrib}
\title{quali_distrib : fonction pour réaliser facilement un grpahique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe}
\usage{
quali_distrib(
  data,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  na.rm = T,
  fill = "deepskyblue3",
  show_n = FALSE,
  show_prop = TRUE,
  prop_method = "beta",
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  labs = TRUE,
  font = "Arial",
  digits = 0,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
quali_distrib : fonction pour réaliser facilement un grpahique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe
}
