% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quali_distrib_group.R
\name{quali_distrib_group}
\alias{quali_distrib_group}
\title{quali_distrib_group : fonction pour calculer facilement des profils lignes par groupe}
\usage{
quali_distrib_group(
  data,
  group,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  caption = NULL,
  show_prop = TRUE,
  unit = "",
  digits = 0,
  dodge = 0.9,
  pretty_pal = "Hokusai1",
  direction = 1,
  wrap_width = 25,
  na.rm = T
)
}
\arguments{
\item{na.rm}{}
}
\description{
quali_distrib_group : fonction pour calculer facilement des profils lignes par groupe
}
