% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runShinyFonctionR}
\alias{runShinyFonctionR}
\title{Shiny fonctionr}
\usage{
runShinyFonctionR(obj)
}
\value{
shinyapp
}
\description{
Shiny fonctionr
}
\examples{
\dontrun{
data(eusilc,data="laeken")
obj <- svydesign(eusilc,ids=~rb030)
runShinyFonctionR(obj)
}
}
