% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_fonctionr}
\alias{theme_fonctionr}
\title{theme_fonctionr}
\usage{
theme_fonctionr(
  font = "Roboto",
  theme = "fonctionr",
  display = "ggplot",
  grid.lines = "x",
  coef_font = 1
)
}
\arguments{
\item{font}{Font used in the graphic. See \code{load_and_active_fonts()} for available fonts.}

\item{theme}{The optionnal theme you want for the graphic. Available themes: \code{"fonctionr"} and \verb{“IWEPS”}. Default is \code{NULL}.}

\item{display}{The way \code{theme_fonctionr()} works on the axis texts: like ggplot2 or ggtext.}

\item{grid.lines}{Specify major grid lines : \code{"x"}, \code{"y"} or \code{"both"}. Default is \code{"x"}.}

\item{coef_font}{A multiplier factor for font size.}
}
\value{
No return value, called for side effects.
}
\description{
A ggplot theme that is ready to use. It is used by most other functions, but can also be applied to an external ggplot object.
}
