% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cat.R
\name{var_cat}
\alias{var_cat}
\title{var_cat : fonction pour réaliser facilement un grpahique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe}
\usage{
var_cat(data, var_cat, reorder = F, na.rm = T, title = "", xlabel = "", ...)
}
\arguments{
\item{...}{}
}
\description{
var_cat : fonction pour réaliser facilement un grpahique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe
}
