% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_group.R
\name{central_group}
\alias{central_group}
\alias{median_group}
\alias{mean_group}
\title{central_group}
\usage{
central_group(
  data,
  group,
  quanti_exp,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  unit = "",
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  xlab = NULL,
  caption = NULL,
  digits = 0,
  show_labs = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  dodge = 0.9,
  reorder = F,
  error_bar = T,
  type,
  fill = NULL,
  na.rm.group = T,
  total_name = "Total",
  font = "Roboto",
  wrap_width = 25,
  export_path = NULL
)

median_group(..., type = "median")

mean_group(..., type = "mean")
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{group}{A variable defining groups be compared.}

\item{quanti_exp}{An expression that define the mean/mediam to be computed.}

\item{facet_var}{A variable defining the faceting group.}

\item{filter_exp}{An expression that filters the data, preserving the design.}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{unit}{Unit showed in the graphic. Default is no unit.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{caption}{Caption of the graphic.}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{show_labs}{TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.}

\item{show_n}{TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.}

\item{show_value}{TRUE if you want to show the mean/median in each group on the graphic. FALSE if you do not want to show the mean/median}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{reorder}{TRUE if you want to reorder the groups according to the mean/median. NA value, in case if na.rm.group = FALSE, is not included in the reorder.}

\item{error_bar}{TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars.}

\item{type}{"mean" to compute means by group ; "median" to compute medians by group.}

\item{fill}{Colour of the bars. NA bar, in case if na.rm.group = FALSE, and total bar are always in grey.}

\item{na.rm.group}{TRUE if you want to remove observations with NA on the group variable or NA on the facet variable. FALSE if you want to create a group with the NA value for the group variable and a facet with the NA value for the facet variable. NA in the variables included in quanti_exp are not affected in this argument. All the observation with a NA in the variables included in quanti_exp are excluded.}

\item{total_name}{Name of the total bar on the graphic. Default is Total.}

\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}

\item{wrap_width}{Number of characters before before going to the line. Applies to the labels of the groups. Default is 25.}

\item{export_path}{Path to export the results in an xlsx file. The file includes two sheets : the table and the graphic.}
}
\value{
A list that contains a table, a graphic and a statistical test
}
\description{
Function to compare means or medians in different groups from complex survey data. It produces a table, a graphic and a statistical test.
}
\examples{
# Loading of data
data(eusilc, package = "laeken")

# Creation of age categories
eusilc$age_cat <- cut(eusilc$age,
breaks = 6,
include.lowest = TRUE)

# Calculation of income means by age category with fonctionr, taking sample design into account
eusilc_mean <- mean_group(
  eusilc,
  group = age_cat,
  quanti_exp = eqIncome / 12,
  strata = db040,
  ids = db030,
  weight = rb050,
  title = "Mean of equivalised income in household by age of individuals",
  subtitle = "Example with austrian SILC data from 'laeken' package"
  )
eusilc_mean$graph

}
