% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/central_group.R
\name{central_group}
\alias{central_group}
\alias{median_group}
\alias{mean_group}
\title{central_group : fonction pour calculer facilement des tendances centrales par groupe}
\usage{
central_group(
  data,
  group,
  quanti_exp,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  unit = "",
  caption = NULL,
  digits = 0,
  show_n = FALSE,
  show_value = TRUE,
  dodge = 0.9,
  reorder = F,
  error_bar = T,
  type,
  fill = NULL,
  na.rm.group = T,
  total_name = "Total",
  font = "Montserrat",
  wrap_width = 25,
  export_path = NULL
)

median_group(..., type = "median")

mean_group(..., type = "mean")
}
\arguments{
\item{export_path}{}
}
\description{
central_group : fonction pour calculer facilement des tendances centrales par groupe
}
