% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_continuous.R,
%   R/distrib_group_continuous.R
\name{distrib_continuous}
\alias{distrib_continuous}
\alias{distrib_group_continuous}
\title{distrib_continuous}
\usage{
distrib_continuous(
  data,
  quanti_exp,
  type = "median",
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.facet = TRUE,
  quantiles = seq(0.1, 0.9, 0.1),
  bw = 1,
  resolution = 1024,
  limits = NULL,
  show_mid_line = TRUE,
  show_ci_lines = TRUE,
  show_ci_area = FALSE,
  show_quant_lines = FALSE,
  show_n = FALSE,
  show_value = TRUE,
  show_lab = TRUE,
  digits = 0,
  unit = "",
  dec = ",",
  pal = c("#00708C", "mediumturquoise"),
  color = NA,
  font = "Roboto",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  export_path = NULL
)

distrib_group_continuous(
  data,
  group,
  quanti_exp,
  type = "median",
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.group = TRUE,
  na.rm.facet = TRUE,
  quantiles = seq(0.1, 0.9, 0.1),
  moustache_probs = c(0.95, 0.8, 0.5),
  bw = 1,
  resolution = 512,
  height = 0.8,
  limits = NULL,
  reorder = FALSE,
  show_mid_point = TRUE,
  show_mid_line = FALSE,
  show_ci_errorbar = TRUE,
  show_ci_lines = FALSE,
  show_ci_area = FALSE,
  show_quant_lines = FALSE,
  show_moustache = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  show_lab = TRUE,
  digits = 0,
  unit = "",
  dec = ",",
  pal = "#e0dfe0",
  pal_moustache = c("#EB9BA0", "#FAD7B1"),
  color = NA,
  alpha = 1,
  font = "Roboto",
  wrap_width_y = 25,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  export_path = NULL
)
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{quanti_exp}{An expression that define the variable the variable to be described.}

\item{type}{Type of central value : "mean" to compute mean as the central value by group ; "median" to compute median as the central value by group..}

\item{facet}{A supprimer?}

\item{filter_exp}{An expression that filters the data, preserving the design.}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{na.rm.facet}{Argument à supprimer?}

\item{quantiles}{Quantiles to be computed in the distributions. Default are deciles.}

\item{bw}{The smoothing bandwidth to be used. The kernels are scaled such that this is the standard deviation of the smoothing kernel. Default is 1.}

\item{resolution}{Resolution of the density curve. Default is 512}

\item{limits}{Limits of the x axe of the graphic. Does not apply to the computation. Default is NULL to show the entire distribution on the graphic.}

\item{show_mid_line}{TRUE if you want to show the mean or median (depending on type) as a line on the graphic. FALSE if you do not want to show it. Default is FALSE}

\item{show_ci_lines}{TRUE if you want to show confidence interval of the mean or median (depending on type) as lines on the graphic. FALSE if you do not want to show it as lines. Default is FALSE}

\item{show_ci_area}{TRUE if you want to show confidence interval of the mean or median (depending on type) as a coloured area on the graphic. FALSE if you do not want to show it as an area. Default is FALSE.}

\item{show_quant_lines}{TRUE if you want to show quantiles as lines on the graphic. FALSE if you do not want to show them as lines. Default is FALSE.}

\item{show_n}{TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.}

\item{show_value}{TRUE if you want to show the mean/median of each group on the graphic. FALSE if you do not want to show the mean/median. Default is TRUE.}

\item{show_lab}{TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.}

\item{digits}{Numbers of digits showed on the value labels on the graphic. Default is 0.}

\item{unit}{Unit showed on the graphic. Default is no unit.}

\item{dec}{Decimal mark shown on the graphic. Default is ",".}

\item{pal}{Color of the density areas. Can be one or sereval colors to create a palette.}

\item{color}{Color of the density curve. Has to be one color.}

\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the X label on the graphic, after the coord_flip(), and not to the x variable in the data. If xlab = NULL, X label on the graphic will be "Moyenne : " + quanti_exp or "Médianne : " + quanti_exp. To show no X label, use xlab = "".}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, ylab refers to the Y label on the graphic, after the coord_flip(), and not to the y variable in the data. If ylab = NULL, Y label on the graphic will be group. To show no Y label, use ylab = "".}

\item{caption}{Caption of the graphic.}

\item{export_path}{Path to export the results in an xlsx file. The file includes three sheets : the table, the graphic and the statistical test.}

\item{group}{A variable defining groups to be compared.}

\item{na.rm.group}{TRUE if you want to remove observations with NA on the group variable. FALSE if you want to create a group with the NA value for the group variable. NA in the variables included in quanti_exp are not affected in this argument. All the observation with a NA in the variables included in quanti_exp are excluded.}

\item{moustache_probs}{A vector defining the proportions of the population used to draw the moustache. Default is 0.95, 0.8, 0.5 to draw a moustache with three groups containing respectively 50 percent, 80 percent and 95 percent of the population closest to the median.}

\item{height}{Height of the curves. Default is 0.8}

\item{reorder}{TRUE if you want to reorder the groups according to the mean/median (depending on type). NA value, in case if na.rm.group = FALSE, is not included in the reorder (A VERIFIER).}

\item{show_mid_point}{TRUE if you want to show the mean or median (depending on type) as a point on the graphic. FALSE if you do not want to show it. Default is TRUE.}

\item{show_ci_errorbar}{TRUE if you want to show confidence interval of the mean or median (depending on type) as an error bar on the graphic. FALSE if you do not want to show it as lines. Default is TRUE.}

\item{show_moustache}{TRUE if you want to show the moustache on the graphic. FALSE if you do not want to show it. Default is TRUE.}

\item{pal_moustache}{Color of the moustache. Can be one or sereval colors to create a palette.}

\item{alpha}{Transparence of the density curve. Default is 1.}

\item{wrap_width_y}{Number of characters before going to the line in the labels of the groups. Default is 25}
}
\value{
A list that contains a table (tab), a graphic (garph) and a density table (dens) and a quantile table (quant)

A list that contains a table, a graphic and a statistical test
}
\description{
Function to describe a continuous variable from complex survey data

Function to compare means or medians in different groups from complex survey data. It produces a table, a graphic and a statistical test.
}
