% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_discrete.R
\name{distrib_discrete}
\alias{distrib_discrete}
\alias{distrib_d}
\title{distrib_discrete}
\usage{
distrib_discrete(
  data,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  na.rm = T,
  fill = "sienna2",
  show_n = FALSE,
  show_value = TRUE,
  reorder = FALSE,
  scale = 100,
  dodge = 0.9,
  prop_method = "beta",
  unit = "\%",
  error_bar = T,
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  show_labs = TRUE,
  font = "Roboto",
  digits = 0,
  wrap_width = 25,
  export_path = NULL
)

distrib_d(...)
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{quali_var}{The discrete variable that is studied.}

\item{facet_var}{A variable defining the faceting group.}

\item{filter_exp}{An expression that filters the data, preserving the design.}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{na.rm}{TRUE if you want to remove the NAs in quali_var. FALSE if you want to create a NA category in the graphic and the table. Default is TRUE.}

\item{fill}{Colour of the bars. NA bar, in case if na.rm.group = FALSE, and total bar are always in grey.}

\item{show_n}{TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.}

\item{show_value}{TRUE if you want to show the proportion of each category on the graphic. FALSE if you do not want to show the proportion.}

\item{reorder}{TRUE if you want to reorder the categories according to their proportion. NA value, in case if na.rm = FALSE, is not included in the reorder.}

\item{scale}{Denominator of the proportion. Default is 100 to interprets numbers as percentages.}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{prop_method}{Type of proportion method to use. See svyciprop in survey package for details. Default is the beta method.}

\item{unit}{Unit showed in the graphic. Default is percent.}

\item{error_bar}{TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars.}

\item{caption}{Caption in the graphic.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{show_labs}{TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.}

\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{wrap_width}{Number of characters before going to the line. Applies to the labels of the categories. Default is 25.}

\item{export_path}{Path to export the results in an xlsx file. The file includes two sheets : the table and the graphic.}
}
\value{
A list that contains a table, a graphic and a statistical test
}
\description{
Function describe the distribution of a discrete variable from complex survey data. It produces a table and a graphic.
}
