% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_discrete.R
\name{distrib_discrete}
\alias{distrib_discrete}
\alias{distrib_d}
\title{distrib_discrete : fonction pour réaliser facilement un graphique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe}
\usage{
distrib_discrete(
  data,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  na.rm = T,
  fill = "sienna2",
  show_n = FALSE,
  show_value = TRUE,
  reorder = FALSE,
  scale = 100,
  dodge = 0.9,
  prop_method = "beta",
  unit = "\%",
  error_bar = T,
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  show_labs = TRUE,
  font = "Roboto",
  digits = 0,
  wrap_width = 25,
  export_path = NULL
)

distrib_d(...)
}
\arguments{
\item{export_path}{}
}
\description{
distrib_discrete : fonction pour réaliser facilement un graphique et une analyse d'une distribution d'une variable qualitative d'un sondage complexe
}
