% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_group_discrete.R
\name{distrib_group_discrete}
\alias{distrib_group_discrete}
\alias{distrib_group_d}
\title{distrib_group_discrete : fonction pour calculer facilement des profils lignes par groupe}
\usage{
distrib_group_discrete(
  data,
  group,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  prop_method = "beta",
  ...,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  show_labs = TRUE,
  show_value = TRUE,
  unit = "",
  digits = 0,
  dodge = 0.9,
  pretty_pal = "Hokusai1",
  direction = 1,
  scale = 100,
  wrap_width = 25,
  legend_ncol = 4,
  font = "Roboto",
  na.rm = T,
  export_path = NULL
)

distrib_group_d(...)
}
\arguments{
\item{na.rm}{}
}
\description{
distrib_group_discrete : fonction pour calculer facilement des profils lignes par groupe
}
