% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esth_graph.R
\name{esth_graph}
\alias{esth_graph}
\title{esth_graph}
\usage{
esth_graph(
  tab,
  ind_var,
  cat_var,
  facet_var = NULL,
  unit = "",
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  scale = 1,
  digits = 2,
  n_var = NULL,
  show_value = TRUE,
  dodge = 0.9,
  reorder = F,
  error_low = NULL,
  error_upp = NULL,
  total_name = NULL,
  fill = "indianred4",
  font = "Roboto",
  wrap_width = 25
)
}
\arguments{
\item{tab}{Table with the variables to be ploted.}

\item{ind_var}{The variable in tab with the indicators to be ploted.}

\item{cat_var}{The variable in tab with the labels of the indicators to be ploted.}

\item{facet_var}{A variable in tab defining the faceting group, if applicable. Default is NULL.}

\item{unit}{The unit showd on the plot. Default is percent.}

\item{caption}{Caption of the graphic.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to cat_var in tab.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, ylab refers to the y label on the graphic, after the coord_flip(), and not to ind_var in tab.}

\item{scale}{Denominator of the proportion. Default is 100 to interprets numbers as percentages.}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{n_var}{The variable in tab containing the number of observation per for each indicator. Default is NULL, not showing the number of observation on the plot.}

\item{show_value}{TRUE if you want to show the values of ind_var on the graphic. FALSE if you do not want to show the proportion. Default is TRUE.}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{reorder}{TRUE if you want to reorder cat_var according to ind_var. FALSE if you do not want to reorder. Default is FALSE.}

\item{error_low}{The variable in tab that is the lower bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.}

\item{error_upp}{The variable in tab that is the upper bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.}

\item{total_name}{Name of the total bar on the graphic.}

\item{fill}{Colour of the bars.}

\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}

\item{wrap_width}{Number of characters before before going to the line. Applies to the labels cat_var. Default is 25.}
}
\description{
Function to construct a graphic following the aestetics of the other function function of this package from a table
}
