% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esth_graph.R
\name{esth_graph}
\alias{esth_graph}
\title{esth_graph}
\usage{
esth_graph(
  tab,
  var,
  value,
  facet = NULL,
  reorder = F,
  error_low = NULL,
  error_upp = NULL,
  pvalue = NULL,
  n_var = NULL,
  show_value = TRUE,
  name_total = NULL,
  scale = 1,
  digits = 2,
  unit = "",
  dec = ",",
  fill = "indianred4",
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL
)
}
\arguments{
\item{tab}{Table with the variables to be ploted.}

\item{var}{The variable in tab with the labels of the indicators to be ploted.}

\item{value}{The variable in tab with the indicators to be ploted.}

\item{facet}{A variable in tab defining the faceting group, if applicable. Default is NULL.}

\item{reorder}{TRUE if you want to reorder var according to value. FALSE if you do not want to reorder. Default is FALSE.}

\item{error_low}{The variable in tab that is the lower bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.}

\item{error_upp}{The variable in tab that is the upper bound of the confidence interval. If either error_low or error_upp is NULL error bars are not shown on the graphic.}

\item{pvalue}{The p-value to show in the caption. It can a numeric value or the pvalue object from a statsistical test.}

\item{n_var}{The variable in tab containing the number of observation per for each indicator. Default is NULL, not showing the number of observation on the plot.}

\item{show_value}{TRUE if you want to show the values of value on the graphic. FALSE if you do not want to show the proportion. Default is TRUE.}

\item{name_total}{Name of the total bar on the graphic.}

\item{scale}{Denominator of the proportion. Default is 100 to interprets numbers as percentages.}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{unit}{The unit showd on the plot. Default is percent.}

\item{dec}{Decimal mark shown on the graphic. Default is ","}

\item{fill}{Colour of the bars.}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{font}{Font used in the graphic. See load_and_active_fonts() for available fonts.}

\item{wrap_width_y}{Number of characters before before going to the line. Applies to the labels var. Default is 25.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to var in tab.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, ylab refers to the y label on the graphic, after the coord_flip(), and not to value in tab.}

\item{caption}{Caption of the graphic.}
}
\description{
Function to construct a graphic following the aestetics of the other function function of this package from a table
}
