% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctionr_options.R
\name{fonctionr_options}
\alias{fonctionr_options}
\title{fonctionr_options}
\usage{
fonctionr_options(
  na.rm.group = NULL,
  na.rm.facet = NULL,
  na.prop = NULL,
  na.vars = NULL,
  na.rm.var = NULL,
  probs = NULL,
  total = NULL,
  prop_method = NULL,
  quantiles = NULL,
  moustache_probs = NULL,
  bw = NULL,
  resolution = NULL,
  height = NULL,
  limits = NULL,
  reorder = NULL,
  position = NULL,
  show_ci = NULL,
  show_mid_point = NULL,
  show_mid_line = NULL,
  show_ci_errorbar = NULL,
  show_ci_lines = NULL,
  show_ci_area = NULL,
  show_quant_lines = NULL,
  show_moustache = NULL,
  show_n = NULL,
  show_value = NULL,
  show_labs = NULL,
  total_name = NULL,
  scale = NULL,
  digits = NULL,
  unit = NULL,
  dec = NULL,
  col = NULL,
  pal = NULL,
  direction = NULL,
  desaturate = NULL,
  lighten = NULL,
  darken = NULL,
  col_density = NULL,
  col_moustache = NULL,
  col_border = NULL,
  alpha = NULL,
  dodge = NULL,
  font = NULL,
  wrap_width_y = NULL,
  wrap_width_leg = NULL,
  legend_ncol = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  lang = NULL,
  theme = NULL,
  coef_font = NULL,
  erase_all = FALSE
)
}
\arguments{
\item{na.rm.group}{na.rm.group}

\item{na.rm.facet}{na.rm.facet}

\item{na.prop}{na.prop}

\item{na.vars}{na.vars}

\item{na.rm.var}{na.rm.var}

\item{probs}{probs}

\item{total}{total}

\item{prop_method}{prop_method}

\item{quantiles}{quantiles}

\item{moustache_probs}{moustache_probs}

\item{bw}{bw}

\item{resolution}{resolution}

\item{height}{height}

\item{limits}{limits}

\item{reorder}{reorder}

\item{position}{position}

\item{show_ci}{show_ci}

\item{show_mid_point}{show_mid_point}

\item{show_mid_line}{show_mid_line}

\item{show_ci_errorbar}{show_ci_errorbar}

\item{show_ci_lines}{show_ci_lines}

\item{show_ci_area}{show_ci_area}

\item{show_quant_lines}{show_quant_lines}

\item{show_moustache}{show_moustache}

\item{show_n}{show_n}

\item{show_value}{show_value}

\item{show_labs}{show_labs}

\item{total_name}{total_name}

\item{scale}{scale}

\item{digits}{digits}

\item{unit}{unit}

\item{dec}{dec}

\item{col}{col}

\item{pal}{pal}

\item{direction}{direction}

\item{desaturate}{desaturate}

\item{lighten}{lighten}

\item{darken}{darken}

\item{col_density}{col_density}

\item{col_moustache}{col_moustache}

\item{col_border}{col_border}

\item{alpha}{alpha}

\item{dodge}{dodge}

\item{font}{font}

\item{wrap_width_y}{wrap_width_y}

\item{wrap_width_leg}{wrap_width_leg}

\item{legend_ncol}{legend_ncol}

\item{title}{title}

\item{subtitle}{subtitle}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{legend_lab}{legend_lab}

\item{caption}{caption}

\item{lang}{lang}

\item{theme}{theme}

\item{coef_font}{coef_font}

\item{erase_all}{TRUE erases all the options. Default is FALSE.}
}
\description{
Function to set global options for fonctionr
}
\examples{
# We set settings for font type and font size
fonctionr_options(font = "Montserrat", coef_font = 1.5)

# Loading of data
data(eusilc, package = "laeken")

# Recoding eusilc$pl030 into eusilc$pl030_rec
eusilc$pl030_rec <- NA
eusilc$pl030_rec[eusilc$pl030 == "1"] <- "Working full time"
eusilc$pl030_rec[eusilc$pl030 == "2"] <- "Working part time"
eusilc$pl030_rec[eusilc$pl030 == "3"] <- "Unemployed"
eusilc$pl030_rec[eusilc$pl030 == "4"] <- "Student"
eusilc$pl030_rec[eusilc$pl030 == "5"] <- "Retired"
eusilc$pl030_rec[eusilc$pl030 == "6"] <- "Permanently disabled"
eusilc$pl030_rec[eusilc$pl030 == "7"] <- "Fulfilling domestic tasks"

# Computation, taking sample design into account
eusilc_prop <- prop_group(
eusilc,
group = pl030_rec,
prop_exp = py090n > 0,
weight = rb050,
title = "\% of ind. receiving unemployment benefits in their hh",
subtitle = "Example with austrian SILC data from 'laeken' package"
)

# Results in graph form
eusilc_prop$graph

}
