% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_and_active_fonts}
\alias{load_and_active_fonts}
\title{load_and_active_fonts}
\usage{
load_and_active_fonts()
}
\description{
Function to load and activate fonctionr's built-in fonts.
Available fonts, included in the package itself, are "Roboto", "Montserrat", "Gotham Narrow", "Helvetica Neue", "League Gothic" and "Amatic". Default is "Roboto".
}
\examples{
# Loading of fonts from fonctionr. You can now use it in fonctionr !
load_and_active_fonts()

}
