% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surface.R
\name{make_surface}
\alias{make_surface}
\title{make_surface}
\usage{
make_surface(
  data,
  IC = TRUE,
  space = NULL,
  reorder = F,
  position = "mid",
  digits = 0,
  unit = NULL,
  pal = "Kandinsky",
  direction = 1,
  last_del = 1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  wrap_lab = 15
)
}
\arguments{
\item{wrap_lab}{}
}
\description{
make_surface
}
