% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_surface.R
\name{make_surface}
\alias{make_surface}
\title{make_surface}
\usage{
make_surface(
  tab,
  var,
  value,
  error_low = NULL,
  error_upp = NULL,
  pvalue = NULL,
  compare = F,
  reorder = F,
  show_ci = TRUE,
  space = NULL,
  position = "mid",
  digits = 0,
  unit = NULL,
  pal = "Kandinsky",
  bg = "snow2",
  direction = 1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  wrap_width_lab = 20
)
}
\arguments{
\item{wrap_width_lab}{}
}
\description{
make_surface
}
