% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_prop.R
\name{many_prop}
\alias{many_prop}
\title{many_prop}
\usage{
many_prop(
  data,
  bin_vars,
  bin_vars_label = NULL,
  facet_var = NULL,
  filter_exp = NULL,
  prop_method = "beta",
  ...,
  unit = "\%",
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  scale = 100,
  digits = 0,
  reorder = FALSE,
  show_labs = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  dodge = 0.9,
  fill = "mediumseagreen",
  error_bar = T,
  font = "Roboto",
  wrap_width = 25
)
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{bin_vars}{A vector containing names of the binarized variables on which to compute the proportions}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{group}{A variable defining groups be compared.}
}
\description{
many_prop
}
