% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_val.R
\name{many_val}
\alias{many_val}
\alias{many_prop}
\alias{many_median}
\alias{many_mean}
\title{many_val}
\usage{
many_val(
  data,
  list_vars,
  type,
  list_vars_lab = NULL,
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.vars = "rm",
  prop_method = "beta",
  reorder = FALSE,
  show_ci = T,
  show_n = FALSE,
  show_value = TRUE,
  show_lab = TRUE,
  scale = NULL,
  digits = 0,
  unit = NULL,
  dec = ",",
  pal = "Egypt",
  direction = 1,
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  export_path = NULL
)

many_prop(..., type = "prop")

many_median(..., type = "median")

many_mean(..., type = "mean")
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{list_vars}{A vector containing names of the dummy variables on which to compute the proportions}

\item{type}{"mean" to compute means ; "median" to compute medians ; "prop" to compute proportions.}

\item{list_vars_lab}{names of the variables}

\item{facet}{A variable defining the faceting group.}

\item{filter_exp}{An expression that filters the data, preserving the design.}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{na.vars}{treatment of NA values in variables. "rm" removes NA only in each individual variable, "rm.all" removes every individual that has at least one NA in one variable.}

\item{prop_method}{Type of proportion method to use. See svyciprop in survey package for details. Default is the beta method.}

\item{reorder}{TRUE if you want to reorder the variables according to the proportion.}

\item{show_ci}{TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars.}

\item{show_n}{TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.}

\item{show_value}{TRUE if you want to show the proportion in each group on the graphic. FALSE if you do not want to show the proportion.}

\item{show_lab}{TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.}

\item{scale}{Denominator of the proportion. Default is 100 to interprets numbers as percentages.}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{unit}{Unit showed in the graphic. Default is percent.}

\item{dec}{Decimal mark shown on the graphic. Default is ","}

\item{pal}{Color palette used on the graphic. The palettes from the packages MetBrewer, MoMAColors and PrettyCols are available.}

\item{direction}{Direction of the palette color. Default is 1. The opposite direction is -1.}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}

\item{wrap_width_y}{Number of characters before before going to the line. Applies to the labels of the groups. Default is 25.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{caption}{Caption of the graphic.}

\item{export_path}{Path to export the results in an xlsx file. The file includes two sheets : the table and the graphic.}
}
\value{
A list that contains a table and a graphic
}
\description{
Function to compute de proportions of a set of several binary variables. It can use complex survey data. It produces a table and a graphic.
}
\examples{
# Loading of data
data(eusilc, package = "laeken")

# Recoding variables
eusilc$worker <- 0
eusilc$worker[eusilc$pl030 == "1"]<-1
eusilc$worker[eusilc$pl030 == "2"]<-1
eusilc$austrian<-0
eusilc$austrian[eusilc$pb220a == "AT"]<-1

# Computation, taking sample design into account
eusilc_many_prop <- many_prop(
eusilc,
list_vars = c(worker,austrian),
list_vars_lab = c("\% of workers","\% of Austrian"),
facet = rb090,
strata = db040,
ids = db030,
weight = rb050,
title = "Proportion of workers and Autrian according to gender",
subtitle = "Example with austrian SILC data from 'laeken' package"
)
eusilc_many_prop$graph
eusilc_many_prop$tab

}
