% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{official_pal}
\alias{official_pal}
\title{official_pal}
\usage{
official_pal(
  inst,
  n,
  direction = 1,
  desaturate = 0,
  lighten = 0,
  darken = 0,
  show_pal = F,
  font = "Gotham Narrow",
  list_pal_names = F
)
}
\arguments{
\item{inst}{name of the institution}

\item{n}{number of colors}

\item{direction}{Direction of the palette color. Default is 1. The opposite direction is -1.}

\item{list_pal_names}{TRUE to generate a vector with palette names.}
}
\description{
Internal function to produce color palettes from different institutions
}
