% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer_survey.R
\name{pivot_longer_survey}
\alias{pivot_longer_survey}
\title{pivot_longer_survey : fonction pour transformer les sorties de srvyr}
\usage{
pivot_longer_survey(data, n_groups)
}
\arguments{
\item{n_groups}{}
}
\description{
pivot_longer_survey : fonction pour transformer les sorties de srvyr
}
