% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_group.R
\name{prop_group}
\alias{prop_group}
\title{prop_group : fonction pour calculer facilement des pourcentages par groupe}
\usage{
prop_group(
  data,
  group,
  prop_exp,
  facet_var = NULL,
  filter_exp = NULL,
  prop_method = "beta",
  ...,
  unit = "\%",
  caption = NULL,
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  xlab = NULL,
  scale = 100,
  digits = 0,
  show_labs = TRUE,
  show_n = FALSE,
  show_value = TRUE,
  dodge = 0.9,
  reorder = F,
  error_bar = T,
  fill = "deepskyblue3",
  na.rm.group = T,
  total_name = "Total",
  font = "Roboto",
  wrap_width = 25,
  export_path = NULL
)
}
\arguments{
\item{export_path}{}
}
\description{
prop_group : fonction pour calculer facilement des pourcentages par groupe
}
