% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_group.R
\name{prop_group}
\alias{prop_group}
\title{prop_group : fonction pour calculer facilement des pourcentages par groupe}
\usage{
prop_group(
  data,
  group,
  expression,
  facet_var = NULL,
  filter_exp = NULL,
  ...,
  unit = "\%",
  caption = NULL,
  scale = 100,
  digits = 1,
  show_n = FALSE,
  dodge = 0.9,
  reorder = F,
  error_bar = T,
  fill = "deepskyblue3",
  na.rm.group = T,
  total_name = "Total",
  wrap_width = 25,
  file_path = NULL
)
}
\arguments{
\item{file_path}{}
}
\description{
prop_group : fonction pour calculer facilement des pourcentages par groupe
}
