% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_group.R
\name{prop_group}
\alias{prop_group}
\title{prop_group}
\usage{
prop_group(
  data,
  group,
  prop_exp,
  group.fill = NULL,
  facet = NULL,
  filter_exp = NULL,
  ...,
  na.rm.group = T,
  na.rm.facet = T,
  na.prop = "rm",
  total = TRUE,
  prop_method = "beta",
  reorder = F,
  show_ci = T,
  show_n = FALSE,
  show_value = TRUE,
  show_labs = TRUE,
  total_name = NULL,
  scale = 100,
  digits = 0,
  unit = "\%",
  dec = NULL,
  pal = NULL,
  direction = 1,
  desaturate = 0,
  lighten = 0,
  darken = 0,
  dodge = 0.9,
  font = "Roboto",
  wrap_width_y = 25,
  wrap_width_leg = 25,
  legend_ncol = 4,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_lab = NULL,
  caption = NULL,
  lang = "fr",
  theme = NULL,
  export_path = NULL
)
}
\arguments{
\item{data}{A dataframe or an object from the survey package or an object from the srvyr package.}

\item{group}{A variable defining the groups to be compared.}

\item{prop_exp}{An expression that define the proportion to be computed.}

\item{group.fill}{A variable defining a second variable of groups to be compared.}

\item{facet}{A variable defining the faceting group.}

\item{filter_exp}{An expression that filters the data, preserving the design.}

\item{...}{All options possible in as_survey_design in srvyr package.}

\item{na.rm.group}{TRUE if you want to remove observations with NA on the group variable or NA on the facet variable. FALSE if you want to create a group with the NA value for the group variable and a facet with the NA value for the facet variable. NA in the variables included in prop_exp are not affected in this argument. All the observation with a NA in the variables included in prop_exp are excluded.}

\item{na.rm.facet}{TRUE if you want to remove observations with NA on the group variable or NA on the facet variable. FALSE if you want to create a group with the NA value for the group variable and a facet with the NA value for the facet variable. NA in the variables included in prop_exp are not affected in this argument. All the observation with a NA in the variables included in prop_exp are excluded.}

\item{na.prop}{"rm" to remove the NA in the variables used in prop_exp before computing the proportions, "include" to compute the proportions with the NA's in the denominators. Default is "rm". When "rm" NA are not allowed in prop_exp.}

\item{total}{TRUE if you want to calculate a total, FALSE if you don't. The default is TRUE}

\item{prop_method}{Type of proportion method to use. See svyciprop in survey package for details. Default is the beta method.}

\item{reorder}{TRUE if you want to reorder the groups according to the proportion. NA value, in case if na.rm.group = FALSE, is not included in the reorder.}

\item{show_ci}{TRUE if you want to show the error bars on the graphic. FALSE if you do not want to show the error bars.}

\item{show_n}{TRUE if you want to show on the graphic the number of individuals in the sample in each group. FALSE if you do not want to show this number. Default is FALSE.}

\item{show_value}{TRUE if you want to show the proportion in each group on the graphic. FALSE if you do not want to show the proportion.}

\item{show_labs}{TRUE if you want to show axes, titles and caption labels. FALSE if you do not want to show any label on axes and titles. Default is TRUE.}

\item{total_name}{Name of the total shown on the graphic. Default is "Total".}

\item{scale}{Denominator of the proportion. Default is 100 to interprets numbers as percentages.}

\item{digits}{Numbers of digits showed on the values labels on the graphic. Default is 0.}

\item{unit}{Unit showed in the graphic. Default is percent.}

\item{dec}{Decimal mark shown on the graphic. Depends on lang: "," for fr and nl ; "." for en.}

\item{pal}{If group.fill is empty, pal must be a vector containing a single color to define the color of the bars. If a variable is specified in group.fill, pal is the color palette used on the graph to differentiate its different modalities. Palettes from fonctionr and the MetBrewer and PrettyCols packages are available. The NA bar, if na.rm.group = FALSE, and the total bar are always in gray.}

\item{direction}{Direction of the palette color. Default is 1. The opposite direction is -1.}

\item{desaturate}{Numeric specifying the amount of desaturation where 1 corresponds to complete desaturation, 0 to no desaturation, and values in between to partial desaturation.}

\item{lighten}{Numeric specifying the amount of lightening. Negative numbers cause darkening.}

\item{darken}{Numeric specifying the amount of lightening. Negative numbers cause lightening.}

\item{dodge}{Width of the bar, between 0 and 1.}

\item{font}{Font used in the graphic. See load_and_active_fonts() for available fonts.}

\item{wrap_width_y}{Number of characters before going to the line. Applies to the labels of the groups. Default is 25.}

\item{wrap_width_leg}{Number of characters before going to the line for the labels of the categories of group.fill. Default is 25.}

\item{legend_ncol}{Number of colomns in the legend. Default is 4.}

\item{title}{Title of the graphic.}

\item{subtitle}{Subtitle of the graphic.}

\item{xlab}{X label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{ylab}{Y label on the graphic. As coord_flip() is used in the graphic, xlab refers to the x label on the graphic, after the coord_flip(), and not to the x variable in the data.}

\item{legend_lab}{Legend (fill) label on the graphic. If legend_lab = NULL, legend label on the graphic will be group.fill. To show no legend label, use legend_lab = "".}

\item{caption}{Caption of the graphic.}

\item{lang}{The language of the indications on the chart. Possibilities: "fr", "nl", "en". Default is "fr".}

\item{theme}{Theme of the graphic. IWEPS adds y axis lines and ticks.}

\item{export_path}{Path to export the results in an xlsx file. The file includes two sheets : the table and the graphic.}
}
\value{
A list that contains a table, a graphic and a statistical test
}
\description{
Function to compare proportions in different groups from complex survey data. It produces a table, a graphic and a statistical test.
}
\examples{
# Loading of data
data(eusilc, package = "laeken")

# Recoding eusilc$pl030 into eusilc$pl030_rec
eusilc$pl030_rec <- NA
eusilc$pl030_rec[eusilc$pl030 == "1"] <- "Working full time"
eusilc$pl030_rec[eusilc$pl030 == "2"] <- "Working part time"
eusilc$pl030_rec[eusilc$pl030 == "3"] <- "Unemployed"
eusilc$pl030_rec[eusilc$pl030 == "4"] <- "Student"
eusilc$pl030_rec[eusilc$pl030 == "5"] <- "Retired"
eusilc$pl030_rec[eusilc$pl030 == "6"] <- "Permanently disabled"
eusilc$pl030_rec[eusilc$pl030 == "7"] <- "Fulfilling domestic tasks"

# Computation, taking sample design into account
eusilc_prop <- prop_group(
eusilc,
group = pl030_rec,
prop_exp = py090n > 0,
strata = db040,
ids = db030,
weight = rb050,
title = "\% of ind. receiving unemployment benefits in their hh",
subtitle = "Example with austrian SILC data from 'laeken' package"
)

# Results in graph form
eusilc_prop$graph

# Results in table format
eusilc_prop$tab

}
