% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quali_distrib_group.R
\name{quali_distrib_group}
\alias{quali_distrib_group}
\title{quali_distrib_group : fonction pour calculer facilement des profils lignes par groupe}
\usage{
quali_distrib_group(
  data,
  group,
  quali_var,
  facet_var = NULL,
  filter_exp = NULL,
  prop_method = "beta",
  ...,
  title = NULL,
  subtitle = NULL,
  ylab = NULL,
  xlab = NULL,
  caption = NULL,
  show_labs = TRUE,
  show_value = TRUE,
  unit = "",
  digits = 0,
  dodge = 0.9,
  pretty_pal = "Hokusai1",
  direction = 1,
  scale = 100,
  wrap_width = 25,
  legend_ncol = 4,
  font = "Roboto",
  na.rm = T,
  export_path = NULL
)
}
\arguments{
\item{na.rm}{}
}
\description{
quali_distrib_group : fonction pour calculer facilement des profils lignes par groupe
}
