% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_fonctionr}
\alias{theme_fonctionr}
\title{theme_fonctionr}
\usage{
theme_fonctionr(font = font)
}
\arguments{
\item{font}{Font used in the graphic. Available fonts, included in the package itself, are "Roboto", "Montserrat" and "Gotham Narrow". Default is "Roboto".}
}
\description{
theme_fonctionr
}
