% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_fonctionr}
\alias{theme_fonctionr}
\title{theme_fonctionr}
\usage{
theme_fonctionr(font = "Roboto")
}
\arguments{
\item{font}{Font used in the graphic. See load_and_active_fonts() for available fonts.}
}
\description{
theme_fonctionr
}
